requirements.txt
Django==1.10.4
gunicorn==19.6.0
psycopg2==2.6.2
Pillow==4.0.0
python-magic==0.4.12
